Attached input action to player; refactored interactable script into interface
