+++ Assets/Scripts/ButtonInteract.cs
@@ -1,0 +1,24 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using UnityEngine.Events;
+
+public class ButtonInteract: MonoBehaviour, IInteractable
+{
+    public UnityEvent action;
+    void Start()
+    {
+        
+    }
+
+    // Update is called once per frame
+    void Update()
+    {
+
+    }
+
+    public void Interact()
+    {
+        action.Invoke();
+    }
+}
