+++ Assets/Scripts/ButtonInteract.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e4aaa058c0b69ab4e9302e11e93fd838
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
