+++ Assets/Scripts/PlayerControls/Interact.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 1fecfa2972ab4614d8e6bdc26de66c48
+guid: 0da2e2d1d30818f428dca5ea0e5ff29e
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
